pipeline mod, unit tests
