+++ serving/src/app/jetstream_proxy.rs
@@ -277,7 +277,7 @@ mod tests {
     use crate::app::callback::store::PayloadToSave;
     use crate::app::callback::CallbackRequest;
     use crate::app::tracker::MessageGraph;
-    use crate::config::pipeline_spec;
+    use crate::pipeline::pipeline_spec;
     use crate::Error;
 
     use super::*;
